feat: add Confirmation to KeyAssignment
This displays a confirmation menu similar to close menu.
If 'yes' is selected, a custom event is emitted which causes the
provided callback to be called.
If 'no' is selected, no further action takes place.
The callback accepts below parameters:
- window (GuiWin)
- pane (MuxPane)
