+++ wezterm-gui/src/overlay/mod.rs
@@ -6,6 +6,7 @@ use std::pin::Pin;
 use std::sync::Arc;
 use wezterm_term::{TerminalConfiguration, TerminalSize};
 
+pub mod confirm;
 pub mod confirm_close_pane;
 pub mod copy;
 pub mod debug;
